* commit '90634f2bc10b5044565e659ab731ba6d937449e4': (84 commits)
  >Cambios trabajo
  Arreglo calculo.xml
  >Precisión campos BBDD
  Repositorios algoritmos desplazamiento
  >Corrección mapper tarea
  >Normalizado BBDD
  > PreLogin Meta4 pool
  Adaptación del modelo, correcciones
  Adaptación del modelo, arreglos
  Adaptación del modelo, tablas:
  Adaptación del modelo, tablas:
  >Meta4 nueva version
  >Meta4 nueva version
  >Nueva version meta4
  >Correccion indice
  >Query algoritmo
  >Adaptación cambios BBDD
  > Framework 2.7.0
  > Añadir control de fecha fin del historico de posiciones
  > Cambios entidades
  ...

+++ APPWEB-INCOMEMotorCalculoJava-api/src/main/java/com/inditex/rrhh/icmclcwb/api/app/calcular/service/AlgoritmoService.java
@@ -14,11 +14,11 @@ public interface AlgoritmoService {
     AlgoritmoDto findByTipoCalculoIdAndTipoComisionId(@NotNull @Positive Long idTipoCalculo,
             @NotNull @Positive Long idTipoComision);
 
-    AlgoritmoDto findById(@NotNull @Positive Long id);
+    AlgoritmoDto findById(@NotNull @Positive Integer id);
 
     List<AlgoritmoDto> findAll();
 
-    List<Long> customFindAlgoritmosIdsByTarea(@NotNull @Positive Long id);
+    List<Integer> customFindAlgoritmosIdsByTarea(@NotNull @Positive Long id);
 
     Boolean checkDuplicatedActives();
